tests/requirements.txt
pytest
pytest-cov
scikit-learn
tensorboard
tensorboardX
colorama
torchvision
pandas
mlflow
omegaconf
matplotlib
pillow
wandb
